+++ bin/generate-tests.php
@@ -47,6 +47,9 @@ $commandFilterList = [
     'parallel:sleep',
     'run:run-parallel',
     'run:ls',
+    // Imported tasks
+    'pyrech:hello',
+    'pyrech:hello-world',
 ];
 $optionFilterList = array_flip(['help', 'quiet', 'verbose', 'version', 'ansi', 'no-ansi', 'no-interaction', 'context']);
 foreach ($applicationDescription['commands'] as $command) {