+++ examples/import.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace import;
+
+use Castor\Attribute\AsTask;
+
+use function Castor\import;
+
+import('github://pyrech/castor-setup-php/main/castor.php');
+
+#[AsTask(description: 'Use a function imported from a remote repository')]
+function hello(): void
+{
+    \pyrech\helloWorld();
+}